Tudo relacionado a árvore funcionando! Compressão completa, leitura da árvore completa.

+++ Huffman/Inc/tree.h
@@ -5,7 +5,6 @@
  * @brief Cria a árvore de huffman. 
  * 
  * @param priority_queue Ponteiro para a fila de prioridade.
- * 
  * @return Retorna um nó que é a raiz da árvore.
 */
 huff_t* create_huff_tree(priority_queue_t* priority_queue);
@@ -21,7 +20,6 @@ void print_preorder(huff_t* root);
  * @brief Diz se a árvore está vazia ou não. 
  * 
  * @param root Ponteiro para a raiz da árvore.
- * 
  * @return Retorna true se a árvore está vazia, ou false se ela não estiver.
  */
 bool is_tree_empty(huff_t* root);
@@ -30,7 +28,6 @@ bool is_tree_empty(huff_t* root);
  * @brief Calcula o tamanho da árvore. 
  * 
  * @param root Ponteiro para a raiz da árvore.
- * 
  * @return Retorna o tamanho da árvore.
  */
-int tree_size(huff_t* root);
+uint16_t tree_size(huff_t* root);
